+++ GoalMogul-iOS/src/Main/Profile/ProfileDetail.js
@@ -1,37 +1,42 @@
 import React, { Component } from 'react';
-import { View, FlatList } from 'react-native';
+import { View, ScrollView } from 'react-native';
+import { connect } from 'react-redux';
 
 /* Component */
-import ProfileDetailCard from './ProfileDetailCard';
+import ProfileDetailCard from './ProfileCard/ProfileDetailCard';
+import ProfileOccupationCard from './ProfileCard/ProfileOccupationCard';
+import ProfileAboutMeCard from './ProfileCard/ProfileAboutMeCard';
+import SearchBarHeader from '../Common/SearchBarHeader';
 
-const testData = [
-  {
-
+const testData = {
+  name: 'Jia Zeng',
+  email: 'jz145@duke.edu',
+  phone: '9194912504',
+  headline: 'I am a student at Duke.',
+  privacy: {
+    friends: 'Public'
+  },
+  profile: {
+    pointsEarned: 10,
+    about: 'This is a test page.',
+    elevatorPitch: 'This is a profile elevator pitch',
+    image: '',
+    occupation: 'Student'
   }
-];
+};
 
 class ProfileDetail extends Component {
 
-  keyExtractor = (item, index) => item.node.image.uri;
-
-  renderRow(data) {
-    // console.log('rendering item p: ', data.item);
-    const p = data.item;
-    return (
-      <ProfileDetailCard />
-    );
-  }
-
   render() {
+    const user = this.props.user;
     return (
       <View style={styles.containerStyle}>
-        <FlatList
-          enableEmptySections
-          data={testData}
-          renderItem={(item) => this.renderRow(item)}
-          numColumns={1}
-          keyExtractor={this.keyExtractor}
-        />
+        <SearchBarHeader backButton />
+        <ScrollView>
+          <ProfileDetailCard data={testData} />
+          <ProfileOccupationCard data={testData} />
+          <ProfileAboutMeCard data={testData} />
+        </ScrollView>
       </View>
     );
   }
@@ -43,4 +48,13 @@ const styles = {
   }
 };
 
-export default ProfileDetail;
+const mapStateToProps = state => {
+  const { userId, user } = state.profile;
+
+  return {
+    userId,
+    user
+  };
+};
+
+export default connect(mapStateToProps, null)(ProfileDetail);