+++ GoalMogul-iOS/src/actions/ProfileActions.js
@@ -1,0 +1,10 @@
+import { Actions } from 'react-native-router-flux';
+
+export const openProfileDetail = () => {
+  return (dispatch) => {
+    dispatch({
+      type: ''
+    });
+    Actions.profileDetail();
+  };
+};