+++ GoalMogul-iOS/src/reducers/index.js
@@ -5,10 +5,12 @@ import AuthReducers from './AuthReducers';
 import RegReducers from './RegReducers';
 import CameraRollReducers from './CameraRollReducers';
 import User from './User';
+import Profile from './Profile';
 
 export default combineReducers({
   auth: AuthReducers,
   registration: RegReducers,
   cameraRoll: CameraRollReducers,
-  user: User
+  user: User,
+  profile: Profile
 });